Fixed an occasional crash in check/dmessage
The private key of a signet signing request is encoded in DER form,
which is a variable-length encoding. In some rare cases the encoded
private key is shorter than 279 bytes. Previously, this has led to
the error message "keys buffer too small for encryption key". This
error message is no longer needed, since the underlying functions
take care of validating the parsing.