Removed an unnecessary else to avoid a GCC warning
When the code is compiled with --coverage, GCC emits one more warning
than without that option. The warning is "assuming signed overflow does
not occur when simplifying conditional to constant".

I suspect that GCC has optimized some code away, and the warning is
only remotely related to what really happens here. In fact, nothing
in this function can possibly generate an integer overflow.